add friendslist component/ separate css files into a styles folder
