+++ src/pages/Dashboard.tsx
@@ -1,5 +1,20 @@
+import FriendsList from '../components/FriendsList';
+import Container from 'react-bootstrap/Container';
+import Row from 'react-bootstrap/Row';
+import Col from 'react-bootstrap/Col';
+import '../styles/Dashboard.css';
+
 const Dashboard = () => {
-  return <h1>The Dashboard</h1>;
+  return (
+    <Container>
+      <Row>
+        <Col>
+          <FriendsList></FriendsList>
+        </Col>
+      </Row>
+    </Container>
+
+  );
 }
 
 export default Dashboard;
