+++ src/styles/Dashboard.css
@@ -1,0 +1,4 @@
+.container {
+  margin: 0;
+  max-width: none;
+}
